pyproject.toml
[tool.poetry]
name = "opencti-attribution-tools"
version = "0.3.2"
description = "OpenCTI STIX2 Attribution Tools Project"
authors = ["Mariam Pogosova <mariam.pogosova@withsecure.com>"]
maintainers = ["WIDR <wit-detection-and-response@withsecure.com>"]
readme = "README.md"
packages = [
    {include = "attribution_tools"},
]

[tool.poetry.dependencies]
python = "^3.8"
pandas = ">=1.0.5"
scikit-learn = ">=1.0.2"
dill = ">=0.3.6"
scipy = ">=1.9.3"
packaging = "^21.3"

[tool.poetry.group.dev.dependencies]
black = "^22.8.0"
isort = "^5.10.1"
mypy = "^0.981"
pytest = "^7.1.3"
pytest-cov = "^3.0.0"
pylint = "^2.15.3"
coverage = "^6.5.0"
pyspark = "^3.3.0"
fs-azkaban-tools = "^3.0.0+1.g1604b69"
jinja2-cli = "^0.8.2"
recommonmark = "^0.7.1"
sphinx = "^5.3.0"
sphinx-markdown-tables = "^0.0.17"
sphinx-rtd-theme = "^1.1.0"
pyyaml = "^6.0"

[tool.poetry.group.emr.dependencies]
boto3 = "1.26.5"

[tool.black]
exclude = ".venv"
line-length = 120
target-version = ['py37', 'py38', 'py39']

[tool.isort]
py_version = "37"
profile = "black"
line_length = 120
force_alphabetical_sort_within_sections = true
include_trailing_comma = true

[tool.mypy]
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "--cov-fail-under=80 --verbose --durations=5 --doctest-modules --maxfail=1 --cov --cov-report term-missing --ignore=docs"

[tool.pylint.'MESSAGES CONTROL']
disable = "too-few-public-methods, too-many-instance-attributes, broad-except, cell-var-from-loop, too-many-locals, f-string-without-interpolation, lost-exception, unspecified-encoding, attribute-defined-outside-init, too-many-arguments"
max-line-length=120

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"